feat(interface): unify onboarding + unlock into one state-agnostic Sign In
