+++ apps/armada-interface/CLAUDE.md
@@ -2,7 +2,7 @@
 
 USDC user app on the Armada protocol — shield, unshield, yield, payments, cross-chain. Replaces the legacy `usdc-v2-frontend` app.
 
-**Status:** V2 shielded-wallet redesign landed (Path C). Deterministic re-sign is the primary unlock path; backup-file + paste-secret remain as secondary recovery for smart-account wallets and cross-device portability. `OnboardingFlowV2` runs welcome → sign → checksum → complete (4 steps); `UnlockFlow` exposes three tabs (Sign in / Backup file / Paste secret). Per-(EVM address, account) walletId map in localStorage drives account-switch detection + auto-lock. Tx history is scoped to the active walletId and AES-256-GCM-encrypted at rest under a per-wallet key. See `specs/TX_SIGNING.md` + `specs/TX_SIGNING_V2_AMENDMENT.md`. Tx flows (shield/unshield/payments/yield) are wired end-to-end.
+**Status:** V2 shielded-wallet redesign landed (Path C). A single state-agnostic **`SignInFlow`** is the account entry for first visit AND returning users — `signIn()` first-derives or re-derives the same identity, so there is no create/unlock split. Backup-file + paste-secret recovery (for smart-account wallets / cross-device) sit behind a "Restore wallet from backup" link; a non-deterministic signer routes to the full-page compatibility screen. Per-(EVM address, account) walletId map in localStorage drives account-switch detection + auto-lock. Tx history is scoped to the active walletId and AES-256-GCM-encrypted at rest under a per-wallet key. See `specs/TX_SIGNING.md` + `specs/TX_SIGNING_V2_AMENDMENT.md`. Tx flows (shield/unshield/payments/yield) are wired end-to-end.
 
 **V1 Phase 9 — chain history recovery (landed):** on unlock, `useHistoryRecovery` reconstructs history from the `@armada/sdk` wallet scan (`runHistoryScan` → `historyEntryToTxRecord`) and synthesizes terminal `TxRecord`s for shields / transacts / unshields / yield ops not already in local IDB. `useIncomingTransferDetector` re-fires the scan on every SDK balance event so received transfers (which the user didn't author) surface live. A per-wallet localStorage checkpoint makes subsequent scans incremental. Settings → "Re-scan history" / "Clear local history" let the user force a refresh or wipe locally. See `lib/shielded/history.ts`, `hooks/useHistoryRecovery.ts`, and `test/integration/history-recovery.test.tsx`.
