+++ apps/armada-interface/src/components/onboarding/SignInFlow.module.css
@@ -1,10 +1,10 @@
-/* ABOUTME: UnlockFlow layout — left-aligned lock icon, body copy, fields, edge-to-edge footer. */
-/* ABOUTME: Visually similar to WelcomeStep but more compact since there's only one field per mode. */
+/* ABOUTME: SignInFlow layout — single sign-in screen (intro + connect/sign) with a quiet restore link; restore view reuses the field/secret styling. */
+/* ABOUTME: Evolved from UnlockFlow.module.css; adds the title/intro + the quiet text-link treatment for restore + back. */
 
 .flow {
   display: flex;
   flex-direction: column;
-  gap: var(--primitives-spacing-11);
+  gap: var(--primitives-spacing-6);
   width: 100%;
   flex: 1;
   min-height: 0;
@@ -15,11 +15,24 @@
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-  gap: var(--primitives-spacing-8);
+  gap: var(--primitives-spacing-6);
   text-align: left;
   width: 100%;
   max-width: 400px;
   margin-inline: auto;
+}
+
+.title {
+  margin: 0;
+  color: var(--semantic-color-text-primary);
+}
+
+.intro {
+  margin: 0;
+  color: var(--semantic-color-text-muted);
+  text-align: left;
+  max-width: 46ch;
+  line-height: var(--primitives-lineHeight-normal);
 }
 
 .modeForm {
@@ -106,9 +119,6 @@
   box-shadow: 0 0 0 2px rgba(196, 145, 229, 0.18);
 }
 
-/* Styled file-input container — replaces the browser's default "Choose file" chrome. The
-   native <input type="file"> sits transparently over the box so the entire area is the click
-   target; the .fileInputDisplay span shows only the selected filename (empty by default). */
 .fileInputArea {
   position: relative;
   display: flex;
@@ -215,16 +225,10 @@
   margin-top: var(--primitives-spacing-6);
 }
 
-.createNew {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: var(--primitives-spacing-1);
-  text-align: center;
-  color: var(--semantic-color-text-muted);
-}
-
-.createNewLink {
+/* Quiet text link — the "Restore wallet from backup" affordance under sign-in, the restore-mode
+   toggle, and the "Back to sign in" link. Left-aligned, understated. */
+.textLink {
+  align-self: flex-start;
   background: none;
   border: 0;
   padding: 0;
@@ -234,7 +238,7 @@
   font: inherit;
 }
 
-.createNewLink:hover,
-.createNewLink:focus-visible {
+.textLink:hover,
+.textLink:focus-visible {
   color: var(--semantic-color-text-primary);
 }
