+++ apps/armada-interface/src/lib/app-mode.test.ts
@@ -1,16 +1,16 @@
 // ABOUTME: Tests for appModeForWalletStatus (P1-14) — the app-mode guard transitions on wallet-status change.
-// ABOUTME: The 'missing' → 'onboarding' case is the fix: Settings → Reset / account-switch-to-unknown must leave app mode instead of rendering a wallet-less shell.
+// ABOUTME: Both 'locked' and 'missing' route to the unified 'signin' screen (sign-in re-derives / first-derives; restore-from-backup is available behind a link).
 
 import { describe, it, expect } from 'vitest'
 import { appModeForWalletStatus } from './app-mode'
 
 describe('appModeForWalletStatus', () => {
-  it('routes a locked wallet back to the unlock screen', () => {
-    expect(appModeForWalletStatus('app', 'locked')).toBe('unlock')
+  it('routes a locked wallet to the sign-in screen', () => {
+    expect(appModeForWalletStatus('app', 'locked')).toBe('signin')
   })
 
-  it('routes a missing wallet to onboarding (Settings → Reset / account-switch-to-unknown)', () => {
-    expect(appModeForWalletStatus('app', 'missing')).toBe('onboarding')
+  it('routes a missing wallet to the sign-in screen (Settings → Reset / account-switch-to-unknown)', () => {
+    expect(appModeForWalletStatus('app', 'missing')).toBe('signin')
   })
 
   it('stays put when the wallet is unlocked', () => {
@@ -18,8 +18,7 @@ describe('appModeForWalletStatus', () => {
   })
 
   it('is a no-op outside app mode — the cold-boot derivation owns the mode there', () => {
-    expect(appModeForWalletStatus('unlock', 'missing')).toBeNull()
-    expect(appModeForWalletStatus('onboarding', 'locked')).toBeNull()
+    expect(appModeForWalletStatus('signin', 'missing')).toBeNull()
     expect(appModeForWalletStatus('pre-init', 'unlocked')).toBeNull()
   })
 })
